Refactor 2FA to server-only and add performance documentation
- Move 2FA library to server-only implementation (2fa-server.ts)
- Update 2FA routes to use server-side library
- Change setup route to return otpauth_url for client-side QR generation
- Remove ProvidePlugin from webpack config for cleaner build
- Add comprehensive performance optimization summary documentation
